requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.3
alpaca-trade-api==3.2.0
requests==2.31.0
